requirements.txt
modelcontextprotocol-sdk
pydantic
python-dotenv
